refactor(enumerators): convert LiquidType to enum class with free functions
- Convert LiquidType from custom class to enum class : uint8_t
- Implement liquid_type_to_string and string_to_liquid_type free functions
- Update all usages in codebase and JSON serialization
- Update Lua bindings to use namespace constants
- Remove obsolete deriveClass for LiquidType

+++ src/creation/liquid.cpp
@@ -66,7 +66,7 @@ json::jnode_t &operator<<(json::jnode_t &lhs, const Liquid &rhs)
 {
     lhs.set_type(json::JTYPE_OBJECT);
     lhs["vnum"] << rhs.vnum;
-    lhs["type"] << rhs.type.toUInt();
+    lhs["type"] << static_cast<uint8_t>(rhs.type);
     lhs["name"] << rhs.name;
     lhs["description"] << rhs.description;
     lhs["flags"] << rhs.flags;
